Struct elsa::map::FrozenBTreeMap

pub struct FrozenBTreeMap<K, V> { /* private fields */ }

Append-only version of std::collections::BTreeMap where insertion does not require mutable access

Implementations

impl<K: Clone + Ord, V: StableDeref> FrozenBTreeMap<K, V>

pub fn new() -> Self

impl<K: Clone + Ord, V: StableDeref> FrozenBTreeMap<K, V>

pub fn insert(&self, k: K, v: V) -> &V::Target

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V::Target>where
K: Borrow<Q>,
Q: Ord,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use elsa::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.get(&2), None);

pub fn map_get<Q: ?Sized, T, F>(&self, k: &Q, f: F) -> Option<T>where
K: Borrow<Q>,
Q: Ord,
F: FnOnce(&V) -> T,

Applies a function to the owner of the value corresponding to the key (if any).

The key may be any borrowed form of the map’s key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use elsa::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.map_get(&1, Clone::clone), Some(Box::new("a")));
assert_eq!(map.map_get(&2, Clone::clone), None);

pub fn into_map(self) -> BTreeMap<K, V>

pub fn as_mut(&mut self) -> &mut BTreeMap<K, V>

Get mutable access to the underlying HashMap.

This is safe, as it requires a &mut self, ensuring nothing is using the ‘frozen’ contents.