elsa::map

Struct FrozenMap

pub struct FrozenMap<K, V, S = RandomState> { /* private fields */ }

Append-only version of std::collections::HashMap where insertion does not require mutable access

Implementations

impl<K: Eq + Hash, V> FrozenMap<K, V>

pub fn new() -> Self

pub fn len(&self) -> usize

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
assert_eq!(map.len(), 0);
map.insert(1, Box::new("a"));
assert_eq!(map.len(), 1);

pub fn is_empty(&self) -> bool

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
assert_eq!(map.is_empty(), true);
map.insert(1, Box::new("a"));
assert_eq!(map.is_empty(), false);

impl<K: Eq + Hash, V: StableDeref, S: BuildHasher> FrozenMap<K, V, S>

pub fn insert(&self, k: K, v: V) -> &V::Target

pub fn get<Q>(&self, k: &Q) -> Option<&V::Target>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the mapโs key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.get(&2), None);

pub fn map_get<Q, T, F>(&self, k: &Q, f: F) -> Option<T>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized, F: FnOnce(&V) -> T,

Applies a function to the owner of the value corresponding to the key (if any).

The key may be any borrowed form of the mapโs key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.map_get(&1, Clone::clone), Some(Box::new("a")));
assert_eq!(map.map_get(&2, Clone::clone), None);

impl<K, V, S> FrozenMap<K, V, S>

pub fn into_tuple_vec(self) -> Vec<(K, V)>

Collects the contents of this map into a vector of tuples.

The order of the entries is as if iterating a HashMap (stochastic).

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
map.insert(1, Box::new("a"));
map.insert(2, Box::new("b"));
let mut tuple_vec = map.into_tuple_vec();
tuple_vec.sort();

assert_eq!(tuple_vec, vec![(1, Box::new("a")), (2, Box::new("b"))]);

pub fn into_map(self) -> HashMap<K, V, S>

impl<K: Eq + Hash + StableDeref, V: StableDeref, S: BuildHasher> FrozenMap<K, V, S>

pub fn get_key_value<Q>(&self, k: &Q) -> Option<(&K::Target, &V::Target)>
where K: Borrow<Q>, Q: Hash + Eq + ?Sized,

Returns a reference to the key and value matching a borrowed key.

The key argument may be any borrowed form of the mapโs key type, but Hash and Eq on the borrowed form must match those for the key type.

Examples

use elsa::FrozenMap;

let map = FrozenMap::new();
map.insert(Box::new("1"), Box::new("a"));
assert_eq!(map.get_key_value(&"1"), Some((&"1", &"a")));
assert_eq!(map.get_key_value(&"2"), None);