Struct elsa::sync::FrozenBTreeMap

pub struct FrozenBTreeMap<K, V>(/* private fields */);

Append-only threadsafe version of std::collections::BTreeMap where insertion does not require mutable access

Implementations

impl<K: Clone + Ord, V: StableDeref> FrozenBTreeMap<K, V>

pub fn new() -> Self

pub fn get<Q>(&self, k: &Q) -> Option<&V::Target>where K: Borrow<Q>, Q: Ord + ?Sized,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but Ord on the borrowed form must match those for the key type.

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.get(&1), Some(&"a"));
assert_eq!(map.get(&2), None);

pub fn insert(&self, k: K, v: V) -> &V::Target

Insert a new value into the map. Does nothing if the key is already occupied.

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.get(&1), Some(&"a"));

pub fn map_get<Q, T, F>(&self, k: &Q, f: F) -> Option<T>where K: Borrow<Q>, Q: Ord + ?Sized, F: FnOnce(&V) -> T,

Applies a function to the owner of the value corresponding to the key (if any).

The key may be any borrowed form of the map’s key type, but Ord on the borrowed form must match those for the key type.

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
assert_eq!(map.map_get(&1, Clone::clone), Some(Box::new("a")));
assert_eq!(map.map_get(&2, Clone::clone), None);

impl<K, V> FrozenBTreeMap<K, V>

pub fn into_tuple_vec(self) -> Vec<(K, V)>

Collects the contents of this map into a vector of tuples.

The order of the entries is as if iterating a BTreeMap (ordered by key).

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
map.insert(1, Box::new("a"));
map.insert(2, Box::new("b"));
let tuple_vec = map.into_tuple_vec();

assert_eq!(tuple_vec, vec![(1, Box::new("a")), (2, Box::new("b"))]);

pub fn len(&self) -> usize

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
assert_eq!(map.len(), 0);
map.insert(1, Box::new("a"));
assert_eq!(map.len(), 1);

pub fn is_empty(&self) -> bool

Examples

use elsa::sync::FrozenBTreeMap;

let map = FrozenBTreeMap::new();
assert_eq!(map.is_empty(), true);
map.insert(1, Box::new("a"));
assert_eq!(map.is_empty(), false);