Struct elsa::sync::LockFreeFrozenVec

pub struct LockFreeFrozenVec<T: Copy> { /* private fields */ }

Append-only threadsafe version of std::vec::Vec where insertion does not require mutable access. Does not lock for reading, only allows Copy types and will spinlock on pushes without affecting reads. Note that this data structure is 64 pointers large on 64 bit systems, in contrast to Vec which is 3 pointers large.

Implementations

impl<T: Copy> LockFreeFrozenVec<T>

pub fn new() -> Self

pub fn push(&self, val: T) -> usize

Pushes an element to the vector, potentially allocating new memory. Returns the index at which the element was inserted.

pub fn get(&self, index: usize) -> Option<T>

Load an element (if it exists). This operation is lock-free and performs minimal synchronization.

pub fn is_empty(&self) -> bool

pub unsafe fn get_unchecked(&self, index: usize) -> T

Load an element (if it exists). This operation is lock-free and performs no synchronized atomic operations. This is a useful primitive to implement your own data structure with newtypes around the index.