Struct elsa::vec::FrozenVec

pub struct FrozenVec<T> { /* private fields */ }

Append-only version of std::vec::Vec where insertion does not require mutable access

Implementations

impl<T> FrozenVec<T>

pub fn new() -> Self

Constructs a new, empty vector.

impl<T: StableDeref> FrozenVec<T>

pub fn push(&self, val: T)

Appends an element to the back of the vector.

pub fn push_get(&self, val: T) -> &T::Target

Push, immediately getting a reference to the element

pub fn get(&self, index: usize) -> Option<&T::Target>

Returns a reference to an element.

pub unsafe fn get_unchecked(&self, index: usize) -> &T::Target

Returns a reference to an element, without doing bounds checking.

Safety

index must be in bounds, i.e. it must be less than self.len()

pub fn len(&self) -> usize

Returns the number of elements in the vector.

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

pub fn first(&self) -> Option<&T::Target>

Returns the first element of the vector, or None if empty.

pub fn last(&self) -> Option<&T::Target>

Returns the last element of the vector, or None if empty.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the vector.

pub fn into_vec(self) -> Vec<T>

Converts the frozen vector into a plain vector.

pub fn as_mut(&mut self) -> &mut Vec<T>

Get mutable access to the underlying vector.

This is safe, as it requires a &mut self, ensuring nothing is using the ‘frozen’ contents.

pub fn binary_search(&self, x: &T::Target) -> Result<usize, usize>where T::Target: Ord,

Binary searches this sorted vector for a given element, analogous to slice::binary_search.

pub fn binary_search_by<'a, F>(&'a self, f: F) -> Result<usize, usize>where F: FnMut(&'a T::Target) -> Ordering,

Binary searches this sorted vector with a comparator function, analogous to slice::binary_search_by.

pub fn binary_search_by_key<'a, B, F>( &'a self, b: &B, f: F ) -> Result<usize, usize>where F: FnMut(&'a T::Target) -> B, B: Ord,

Binary searches this sorted vector with a key extraction function, analogous to slice::binary_search_by_key.

pub fn partition_point<P>(&self, pred: P) -> usizewhere P: FnMut(&T::Target) -> bool,

Returns the index of the partition point according to the given predicate (the index of the first element of the second partition), analogous to slice::partition_point.